other/backend/requirements.txt
fastapi
uvicorn
web3
pyjwt
qrcode
Pillow
python-multipart
mnemonic
langchain-groq
python-dotenv
langchain-community
langchain-classic
bcrypt
pyotp
